Fix gcc compiler warning about unsigned/signed integer comparison
gcc 4.4.3 with option "-Wall" complained about
  'comparison between signed and unsigned integer expressons'
in every code path using testFlag().

+++ src/corelib/global/qglobal.h
@@ -2170,7 +2170,7 @@ public:
 
     inline bool operator!() const { return !i; }
 
-    inline bool testFlag(Enum f) const { return (i & f) == f && (f != 0 || i == f ); }
+    inline bool testFlag(Enum f) const { return (i & f) == f && (f != 0 || i == (int)f ); }
 };
 
 #define Q_DECLARE_FLAGS(Flags, Enum)\
